requirements.txt
python-telegram-bot==20.7
requests==2.31.0
firebase-admin==6.4.0
python-dotenv==1.0.1
openai==1.12.0
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1
joblib==1.3.2
matplotlib==3.8.3
gunicorn==21.2.0
aiohttp==3.9.3